Update 2.0
The visualize_referral_count function includes the platform parameter to handle data in different formats that may be returned by the API for iOS and Android apps. When using this function, you must specify the platform ('ios' or 'android') and provide the appropriate API URL for referral count data. The function will then make a request to this URL to retrieve the data and generate a histogram plot.